error fixes

+++ src/main/java/com/ecobank/intern_portal/controller/TaskController.java
@@ -5,8 +5,8 @@ import com.ecobank.intern_portal.model.Task;
 import com.ecobank.intern_portal.repository.TaskRespository;
 import com.ecobank.intern_portal.service.TaskService;
 import lombok.AllArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
+import org.springframework.http.*;
+import org.springframework.web.ErrorResponse;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -34,6 +34,10 @@ public class TaskController {
 
     @GetMapping("/tasks/{id}")
     public ResponseEntity<TaskDto> getTaskById(@PathVariable("id") Long id) {
+        if(taskService.getTaskById(id) == null){
+            System.out.println("Task not found");
+            return ResponseEntity.notFound().build();
+        }
         TaskDto taskById = taskService.getTaskById(id);
         return ResponseEntity.ok(taskById);
     }
@@ -46,6 +50,21 @@ public class TaskController {
 
     @DeleteMapping("/tasks/{id}")
     public ResponseEntity<Void> deleteTask(@PathVariable("id") Long id) {
+        TaskDto taskById = taskService.getTaskById(id);
+        if(taskById == null){
+            ErrorResponse errorResponse = new ErrorResponse() {
+                @Override
+                public HttpStatusCode getStatusCode() {
+                    return null;
+                }
+
+                @Override
+                public ProblemDetail getBody() {
+                    return null;
+                }
+            };
+            return ResponseEntity.notFound().build();
+        }
         taskService.deleteTask(id);
         return ResponseEntity.noContent().build();
     }
